Cleaned up wrapping
Previously, the straight-forward use of Text::Wrap::wrap() on input
which contained a comment starting midway through a line would result
in that line being over-long. This update fixes that by subtracting
the length of the last line of output from the first line of the
comment and stitching them together.